WinMD: add some decoding helpers for `CodedIndex`
A `CodedIndex` encodes a discriminator in the low log(n) bits of the
`CodedIndex` and the row index in the high (16 - log(n) bits if the
tables are small or 32 - log(n) bits if the tables are large).  We
compute the mask for the discriminator by computing the number of bits
required to discriminate between the tables associated with the
`CodedIndex` and use that to extract the tag and the row index.