WinMD: add table iteration support
Change the generic table implementation to a more concrete implementation by
removing the conformance requirement.  This requires changing the tuple based
table schema definition to an array one.  However, it allows for a more uniform
iteration of the tables.  Introduce the new `DatabaseCompressionState` which
provides the context for the decompression of the table.  With this, we are able
to iterate the records of the table via the new `Records` type.  The reason for
`Table` not being directly iterable is that the sequence conformance would
introduce a `Self` requirement, which restricts the ability to treat all the
tables identically.  We now truncate the data slice that the table references
allowing catching the out of bounds accesses earlier.