winmd-inspect: clean up structure of the command
The inspection tool is meant to be a "swiss army knife" for WinMD files.
It should allow you to inspect various aspects (e.g. identify all the
types contained within the WinMD "database") in addition to dump.
Change the tool to use the sub-command style, providing the following
(hypothetical) interface:

~~~
> winmd-inspect Win32.winmd dump
...
> winmd-inspect Win32.winmd typdefs
...
> winmd-inspect Win32.winmd types
...
~~~