WinMD: be more careful about failures in accessing data
This adds more checking to the data handling for reading tables.
Because the input is technically untrusted, we should be more careful
about invalid accesses.  This relies on newer language features, but
given that we are still in the process of designing this library, it is
safe to assume that we can use nearly the latest release of the
toolchain.