WinMD: re-implement iteration support
This massively cleans up the iteration support in WinMD by rethinking through
the implementation.  It massively simplifies the iteration process by hoisting
the `RecordReader` constructor into `Database`.  Instead a new `TableIterator`
type implements the `IteratorProtocol` for a given `Table`.  By passing along
the type information for the table, we can hoist the table processing operation
into the iterator element type construction (i.e. the `next` operation).  This
allows us to strip the element down to a pre-processed array of `Int` and the
heap references.  The element itself can also preserve type information,
allowing type-safe access to the fields on a per-table basis.