WinMD: replace heap refs with database in iteration
Provide a reference to the database to the iterator.  This is needed
since the iterator requires the ability to inspect not only the current
table but other tables in the database in order to compute runs which
have have a foreign key constraint that needs to inspect the maximal
count for the foreign table.